requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
amqp==5.1.1; python_version >= '3.6'
asgiref==3.5.2; python_version >= '3.7'
attrs==22.1.0; python_version >= '3.5'
billiard==3.6.4.0
celery==5.2.7
certifi==2022.6.15.2; python_version >= '3.6'
cffi==1.15.1
charset-normalizer==2.1.1; python_version >= '3.6'
click-didyoumean==0.3.0; python_full_version >= '3.6.2' and python_full_version < '4.0.0'
click-plugins==1.1.1
click-repl==0.2.0
click==8.1.3; python_version >= '3.7'
coreapi==2.3.3
coreschema==0.0.4
django-debug-toolbar==3.6.0
django-rest-swagger==2.2.0
django==4.1.1
djangorestframework-simplejwt==5.2.0
djangorestframework==3.13.1
idna==3.4; python_version >= '3.5'
iniconfig==1.1.1
itypes==1.2.0
jinja2==3.1.2; python_version >= '3.7'
kombu==5.2.4; python_version >= '3.7'
markupsafe==2.1.1; python_version >= '3.7'
openapi-codec==1.3.2
packaging==21.3; python_version >= '3.6'
pluggy==1.0.0; python_version >= '3.6'
prompt-toolkit==3.0.31; python_full_version >= '3.6.2'
psycopg2-binary==2.9.3
py==1.11.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
pycparser==2.21
pyjwt==2.4.0; python_version >= '3.6'
pyparsing==3.0.9; python_full_version >= '3.6.8'
pytest==7.1.3
pytz==2022.2.1
rabbitmq==0.2.0
requests==2.28.1
simplejson==3.17.6; python_version >= '2.5' and python_version not in '3.0, 3.1, 3.2, 3.3'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.2; python_version >= '3.5'
tomli==2.0.1; python_version >= '3.7'
uritemplate==4.1.1; python_version >= '3.6'
urllib3==1.26.12; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5' and python_full_version < '4.0.0'
vine==5.0.0; python_version >= '3.6'
wcwidth==0.2.5
